move formValid to /server/middlewares
